+++ skoolkit/skoolmacro.py
@@ -560,6 +560,24 @@ def parse_chr(writer, text, index, *cwd):
     end, num = parse_ints(text, index, 1, fields=writer.fields)
     return end, writer.to_chr(num)
 
+def parse_copy(text, index, fields, frame_map=None):
+    # #COPY[x,y,width,height][{CROP}](old,new)
+    defaults = (0, 0, None, None)
+    try:
+        index, x, y, width, height = parse_ints(text, index, 4, defaults, ('x', 'y', 'width', 'height'), fields)
+    except InvalidParameterError:
+        x, y, width, height = defaults
+    if index < len(text) and text[index] == '{':
+        end, crop_rect = _parse_crop_spec(text, index, fields)
+    else:
+        end, crop_rect = index, None
+    end, (old, new) = parse_strings(text, end, 2)
+    if frame_map is not None:
+        if old not in frame_map:
+            raise MacroParsingError('No such frame: "{}"'.format(old))
+        frame_map[new] = frame_map[old].copy(new, x, y, width, height, crop_rect)
+    return end, ''
+
 def parse_d(writer, text, index, *cwd):
     # #Daddr
     end, addr = parse_ints(text, index, 1, fields=writer.fields)