+++ sphinx/source/changelog.rst
@@ -5,6 +5,8 @@ Changelog
 -----
 * Dropped support for Python 3.5
 * Added the :ref:`OVER` macro (for superimposing one frame on another)
+* Added the :ref:`COPY` macro (for copying all or part of an existing frame
+  into a new frame)
 * Added the ``defaults``, ``dsvalue`` and ``dvalues`` parameters to the
   :ref:`DEFINE` macro (to enable an optional string parameter and optional
   integer parameters for the defined macro and to specify their default values)